+++ zcash_primitives/src/transaction/builder.rs
@@ -552,7 +552,7 @@ mod tests {
         legacy::TransparentAddress,
         memo::MemoBytes,
         merkle_tree::{CommitmentTree, IncrementalWitness},
-        sapling::Rseed,
+        sapling::{Node, Rseed},
         transaction::components::{
             amount::{Amount, DEFAULT_FEE},
             sapling::builder::{self as build_s},
@@ -671,7 +671,7 @@ mod tests {
         let mut rng = OsRng;
 
         let note1 = to.create_note(50000, Rseed::BeforeZip212(jubjub::Fr::random(&mut rng)));
-        let cmu1 = note1.commitment();
+        let cmu1 = Node::from_cmu(&note1.cmu());
         let mut tree = CommitmentTree::empty();
         tree.append(cmu1).unwrap();
         let witness1 = IncrementalWitness::from_tree(&tree);
@@ -779,7 +779,7 @@ mod tests {
         }
 
         let note1 = to.create_note(50999, Rseed::BeforeZip212(jubjub::Fr::random(&mut rng)));
-        let cmu1 = note1.commitment();
+        let cmu1 = Node::from_cmu(&note1.cmu());
         let mut tree = CommitmentTree::empty();
         tree.append(cmu1).unwrap();
         let mut witness1 = IncrementalWitness::from_tree(&tree);
@@ -817,7 +817,7 @@ mod tests {
         }
 
         let note2 = to.create_note(1, Rseed::BeforeZip212(jubjub::Fr::random(&mut rng)));
-        let cmu2 = note2.commitment();
+        let cmu2 = Node::from_cmu(&note2.cmu());
         tree.append(cmu2).unwrap();
         witness1.append(cmu2).unwrap();
         let witness2 = IncrementalWitness::from_tree(&tree);